Added some missing input fields
I had forgotten to include fields to record the type of detection and age of the bird.

I also created a script to help generate observation data files from CSV files from prior years.
Unfortunately, most data only exists as photos of paper forms, so creating the CSV files will
require a lot of work. But once that's done, it's now easy to generate the observation files
and upload the data.