User can select year to map
Modified the map page to plot the year specified in a dropdown
(defaults to the most recent year for which data is available).

Also modified the data page to only regenerate the data summary CSV files
if observation files have been added or changed since the summary files were
last generated.