Convert newlines and other control chars in the notes field to the carat symbol: "^".
Leaving in control chars such as newlines breaks the code that passes summary data to
the jinja templates.  Taking them out makes the notes a bit harder to read, but
in most cases it should be easy to convert back for display by treating a carat as a
newline.

Also fixed a bug in the code to display the map when there are no data points for the
current year (as will typically be true from January to the end of March).